Update jogl jars to latest version of jogl, so GL demos could work on Mac. Modify JaneliaDebugGL2 to work with latest JOGL. SliceViewer changes so TracedPathSegments render on Mac. Use int MNEMONIC_KEY in SliceViewer actions to avoid weird error on Mac Mountain Lion. Refine OpenGL demos to run on Mac.
Former-commit-id: ff5868d0cf2975ada12cee74a697c937c6c98be3

+++ console/src/main/java/org/janelia/it/FlyWorkstation/gui/slice_viewer/OrthogonalViewer.java
@@ -456,7 +456,7 @@ implements MouseModalWidget, TileConsumer
 			super(-1, viewer);
 			putValue(NAME, "Previous "+viewer.sliceAxis.getName()+" Slice");
 			putValue(SMALL_ICON, Icons.getIcon("z_stack_up.png"));
-			putValue(MNEMONIC_KEY, KeyEvent.VK_PAGE_UP);
+			putValue(MNEMONIC_KEY, (int)KeyEvent.VK_PAGE_UP);
 			KeyStroke accelerator = KeyStroke.getKeyStroke(
 				KeyEvent.VK_PAGE_UP, 0);
 			putValue(ACCELERATOR_KEY, accelerator);
@@ -472,7 +472,7 @@ implements MouseModalWidget, TileConsumer
 			super(1, viewer);
 			putValue(NAME, "Next "+viewer.sliceAxis.getName()+" Slice");
 			putValue(SMALL_ICON, Icons.getIcon("z_stack_down.png"));
-			putValue(MNEMONIC_KEY, KeyEvent.VK_PAGE_DOWN);
+			putValue(MNEMONIC_KEY, (int)KeyEvent.VK_PAGE_DOWN);
 			KeyStroke accelerator = KeyStroke.getKeyStroke(
 				KeyEvent.VK_PAGE_DOWN, 0);
 			putValue(ACCELERATOR_KEY, accelerator);
